requirements.txt
praw
openai
python-dotenv
tqdm
groq
